(build) Create a structure for the project and a Makefile structure for the compilation and flashing process.
Create a base folder structure that will be filled with files and additional folders as the project grows.
Creating a Makefile structure for complication and flashing of the STM32F3 microcontroller.
Creating a temporary configuration of clocks and GPIOs for the LED for compile check, flashing and debugging.
Adding a README.md file that will contain the most important information about the project: the software needed to build and flash, the project structure, architecture and schematics, etc.
Adding a .gitignore file in which folders where files created during compilation as well as files from the .vscode folder were added